requirements.txt
elasticsearch>=8.0.0
faiss-cpu>=1.7.0
numpy>=1.21.0
torch>=2.0.0
transformers>=4.30.0
python-dotenv>=0.19.0
tqdm>=4.62.0
torch-geometric>=2.3.0
spacy>=3.6.0
flask>=2.3.0
fastapi>=0.100.0
uvicorn>=0.22.0
langchain>=0.0.267
pandas>=2.0.0
scikit-learn>=1.2.0 